fix(gate): extract check 7 to helper to stay under 800-line architecture limit
Gatekeeper.sh Python architecture guard rejects files over 800 lines.
PR #1488 added check 7 (hollow-compliance validation) which pushed
ai_workflow_gate.py from 759 to 910 lines, triggering the guard.

Fix: extract all check 7 constants, helper functions, and the
check_section_content_quality() callable into a new helper module
at scripts/ops/helpers/section_content_quality.py.

Main gate file: 910 → 784 lines (under 800 limit).

Also fix test _check_content() wrapper to correctly pass the
section_text_between callback with proper arg ordering (the helper
calls fn(heading, body) but gate.section_text_between expects
(body, heading) — validate() already used a lambda for this but
the tests were passing the raw function).
